Refactor de_export.py, extract template functions
Start moving functionality for amici model code generation to a private
subpackage `amici._codegen`. More to follow.
